Move views/exceptions.py to views/errors.py
These are called "error views" not "exception views" in Pyramid, and
this makes way for views/exceptions.py to contain exception classes
used by views like other */exceptions.py files do.
