+++ tests/h/views/errors_test.py
@@ -4,7 +4,7 @@ from __future__ import unicode_literals
 
 from mock import Mock
 
-from h.views.exceptions import notfound, error, json_error
+from h.views.errors import notfound, error, json_error
 
 
 def test_notfound_view(pyramid_request):
@@ -15,7 +15,7 @@ def test_notfound_view(pyramid_request):
 
 
 def test_error_view(patch, pyramid_request):
-    handle_exception = patch("h.views.exceptions.handle_exception")
+    handle_exception = patch("h.views.errors.handle_exception")
     exception = Mock()
 
     result = error(exception, pyramid_request)
@@ -25,7 +25,7 @@ def test_error_view(patch, pyramid_request):
 
 
 def test_json_error_view(patch, pyramid_request):
-    handle_exception = patch("h.views.exceptions.handle_exception")
+    handle_exception = patch("h.views.errors.handle_exception")
     exception = Mock()
 
     result = json_error(exception, pyramid_request)
